Refactor Survey Service and Controller into CRUD Services
