+++ src/main/java/de/fsr/mariokart_backend/survey/service/SurveyUpdateService.java
@@ -1,22 +1,15 @@
 package de.fsr.mariokart_backend.survey.service;
 
-import java.util.List;
-
 import org.springframework.stereotype.Service;
 
 import de.fsr.mariokart_backend.survey.model.Question;
-import de.fsr.mariokart_backend.survey.model.dto.AnswerInputDTO;
-import de.fsr.mariokart_backend.survey.model.dto.AnswerReturnDTO;
 import de.fsr.mariokart_backend.survey.model.dto.QuestionInputDTO;
 import de.fsr.mariokart_backend.survey.model.dto.QuestionReturnDTO;
 import de.fsr.mariokart_backend.survey.model.subclasses.CheckboxQuestion;
 import de.fsr.mariokart_backend.survey.model.subclasses.FreeTextQuestion;
 import de.fsr.mariokart_backend.survey.model.subclasses.MultipleChoiceQuestion;
 import de.fsr.mariokart_backend.survey.model.subclasses.TeamQuestion;
-import de.fsr.mariokart_backend.survey.repository.AnswerRepository;
 import de.fsr.mariokart_backend.survey.repository.QuestionRepository;
-import de.fsr.mariokart_backend.survey.service.dto.AnswerInputDTOService;
-import de.fsr.mariokart_backend.survey.service.dto.AnswerReturnDTOService;
 import de.fsr.mariokart_backend.survey.service.dto.QuestionInputDTOService;
 import de.fsr.mariokart_backend.survey.service.dto.QuestionReturnDTOService;
 import jakarta.persistence.EntityNotFoundException;
@@ -24,56 +17,10 @@ import lombok.AllArgsConstructor;
 
 @Service
 @AllArgsConstructor
-public class SurveyService {
-
+public class SurveyUpdateService {
     private final QuestionRepository questionRepository;
-    private final AnswerRepository answerRepository;
     private final QuestionInputDTOService questionInputDTOService;
     private final QuestionReturnDTOService questionReturnDTOService;
-    private final AnswerInputDTOService answerInputDTOService;
-    private final AnswerReturnDTOService answerReturnDTOService;
-
-    public QuestionReturnDTO getQuestion(Long id) {
-        return questionReturnDTOService.questionToQuestionReturnDTO(questionRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("There is no question with this id.")));
-    }
-
-    public List<QuestionReturnDTO> getQuestions() {
-        return questionRepository.findAll().stream()
-                .map(questionReturnDTOService::questionToQuestionReturnDTO)
-                .toList();
-    }
-
-    public List<QuestionReturnDTO> getVisibleQuestions() {
-        return questionRepository.findAllByVisible(true).stream()
-                .map(questionReturnDTOService::questionToQuestionReturnDTO)
-                .toList();
-    }
-
-    public List<AnswerReturnDTO> getAnswersOfQuestion(Long id) {
-        return answerRepository.findAllByQuestionId(id).stream()
-                .map(answerReturnDTOService::answerToAnswerReturnDTO)
-                .toList();
-    }
-
-    public QuestionReturnDTO createSurvey(QuestionInputDTO survey) {
-        // return
-        // questionRepository.save(questionInputDTOService.questionInputDTOToQuestion(survey));
-        return questionReturnDTOService.questionToQuestionReturnDTO(
-                questionRepository.save(questionInputDTOService.questionInputDTOToQuestion(survey)));
-    }
-
-    public AnswerReturnDTO submitAnswer(AnswerInputDTO answer) {
-        // return
-        // answerRepository.save(answerInputDTOService.answerInputDTOToAnswer(answer));
-        Question question = questionRepository.findById(answer.getQuestionId())
-                .orElseThrow(() -> new EntityNotFoundException("There is no question with this id."));
-        if (!question.getActive() || !question.getVisible()) {
-            throw new IllegalArgumentException("Question is not active or visible.");
-        }
-        return answerReturnDTOService
-                .answerToAnswerReturnDTO(answerRepository.save(answerInputDTOService.answerInputDTOToAnswer(answer)));
-    }
 
     public QuestionReturnDTO updateQuestion(Long id, QuestionInputDTO question) {
         Question questionToUpdate = questionRepository.findById(id)
@@ -115,14 +62,4 @@ public class SurveyService {
         }
         return questionReturnDTOService.questionToQuestionReturnDTO(questionRepository.save(questionToUpdate));
     }
-
-    public void deleteQuestion(Long id) {
-        questionRepository.deleteById(id);
-        answerRepository.deleteAllByQuestionId(id);
-    }
-
-    public void deleteAllQuestions() {
-        questionRepository.deleteAll();
-        answerRepository.deleteAll();
-    }
-}
+}
